requirements.txt
termcolor
pillow
tqdm